pyproject.toml
[project]
name = "ncc-dashboard-bc"
version = "0.1.0"
description = ""
authors = [
    {name = "officialalkenes",email = "bello.abdul@ivendigital.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "python-decouple (>=3.8,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
